prefix command functions with group

+++ rcbot.py
@@ -4,6 +4,7 @@ import typing
 import data
 import datetime
 import Levenshtein as lev
+import os
 
 # client = discord.Client()
 bot = commands.Bot(command_prefix='?')
@@ -15,6 +16,51 @@ global_ = Globals()
 global_.platform = data.Platform()
 global_.branch = -1
 global_.section = -1
+global_.int_conversion_error = False
+
+
+async def ensure_int(val: typing.Any, ctx = -1) -> bool:
+    try:
+        int(val)
+        return True
+    except ValueError:
+        if ctx != -1:
+            await ctx.send(f"`{val}` is not an integer")
+        return False
+
+def ensure_ints(*val: typing.Any) -> bool:
+    out = True
+    for v in val:
+        out &= ensure_int(v)
+    return out
+
+
+# class Int__AlertInChat:
+#     def __init__(self, message = ""):
+#         self.message = message
+
+#     @classmethod
+#     async def convert(cls, ctx, argument):
+#         if not ensure_int(argument):
+#             # await ctx.send(self.message)
+#             global_.int_conversion_error = True
+#         else:
+#             global_.int_conversion_error = False
+#         return argument
+
+
+class Int__AlertInChat(commands.Converter):
+    def __init__(self, message = "Int not provided for int field"):
+        self.message = message
+
+    async def convert(self, ctx, argument):
+        if not await ensure_int(argument, ctx):
+            await ctx.send(self.message)
+            global_.int_conversion_error = True
+        else:
+            global_.int_conversion_error = False
+        return argument
+
 
 @bot.event
 async def on_ready() -> None:
@@ -40,8 +86,9 @@ async def branch(ctx):
         await ctx.send('Invalid subcommand.')
 
 @branch.command(name="fork")
-async def fork(ctx, branch_id: int = -1, name: str = ""):
-    if branch_id == -1:
+async def branch_fork(ctx, branch_id: Int__AlertInChat("Invalid branch id"), name: str = ""):
+    if global_.int_conversion_error:
+        print("Bad integer input")
         await ctx.send("Invalid branch id, appending new branch.")
     if name == "":
         current_datetime = str(datetime.datetime.now())
@@ -53,36 +100,43 @@ async def fork(ctx, branch_id: int = -1, name: str = ""):
     await ctx.send(f"New branch `{name}` id `{global_.platform.fork(branch_id, name)}`.")
 
 @branch.command(name="display")
-async def display(ctx, branch_id: int = -1):
-    if 0 <= branch_id <= 1:
-        await ctx.send(f"Branch {branch_id} `{global_.platform[branch_id].name}` (protected) - \
-            {len(global_.platform[branch_id].sections)} sections.")
-    elif branch_id == -1:
+async def branch_display(ctx, branch_id: Int__AlertInChat("Invalid branch id")):
+    if global_.int_conversion_error:
+        print("Bad integer input")
         first_line = f"Listing {len(global_.platform)} branches"
         line_limit_exceeded = ", limited to first 30" if len(global_.platform) > 30 else ''
         line = "-" * int((len(first_line) + len(line_limit_exceeded)) * 1.4)
         await ctx.send(first_line + line_limit_exceeded + f":\n{line}\n" + "\n".join([f"Branch {i} `{x.name}` - \
             {len(x.sections)} sections" for i, x in zip(range(30), global_.platform)]) + ".")
+    if 0 <= branch_id <= 1:
+        await ctx.send(f"Branch {branch_id} `{global_.platform[branch_id].name}` (protected) - \
+            {len(global_.platform[branch_id].sections)} sections.")
     else:
         await ctx.send(f"Branch {branch_id} `{global_.platform[branch_id].name}` - \
             {len(global_.platform[branch_id].sections)} sections.")
 
 @branch.command(name="count")
-async def count(ctx):
+async def branch_count(ctx):
     await ctx.send(f"{len(global_.platform)} branches.")
 
 @branch.command(name="select")
-async def select(ctx, branch_id: int = -1):
+async def branch_select(ctx, branch_id: Int__AlertInChat("Invalid branch id")):
+    if global_.int_conversion_error:
+        print("Bad integer input, returning")
+        return
     global_.branch = branch_id
     await ctx.send(f"Set branch selector to `{branch_id}`.")
 
 @branch.command(name="selected")
-async def selected(ctx):
+async def branch_selected(ctx):
     await ctx.send(f"Branch selector is `{global_.branch}`.")
 
 @branch.command(name="delete")
-async def delete(ctx, branch_id: int = -1):
-    global_.branch[branch_id].delete()
+async def branch_delete(ctx, branch_id: Int__AlertInChat("Invalid branch id")):
+    if global_.int_conversion_error:
+        print("Bad integer input, returning")
+        return
+    global_.platform[branch_id].delete()
     await ctx.send(f"Branch `{branch_id}` deleted.")
 
 
@@ -92,8 +146,9 @@ async def section(ctx):
         await ctx.send('Invalid subcommand.')
 
 @section.command(name="add")
-async def add(ctx, section_id: int = -1, name: str = ""):
-    if section_id == -1:
+async def section_add(ctx, section_id: Int__AlertInChat("Invalid section id"), name: str = ""):
+    if global_.int_conversion_error:
+        print("Bad integer input")
         await ctx.send("Invalid section id, appending new section.")
     if name == "":
         current_datetime = str(datetime.datetime.now())
@@ -105,18 +160,25 @@ async def add(ctx, section_id: int = -1, name: str = ""):
     await ctx.send(f"New section `{name}` id `{global_.platform[global_.branch].add(name, section_id)}`.")
 
 @section.command(name="swap")
-async def swap(ctx, section_id_1: int, section_id_2: int):
+async def section_swap(ctx, section_id_1: Int__AlertInChat("Invalid section id"), section_id_2: Int__AlertInChat("Invalid section id")):
+    if global_.int_conversion_error:
+        print("Bad integer input, returning")
+        return
     global_.platform[global_.branch].swap(section_id_1, section_id_2)
     await ctx.send(f"Branch id `{section_id_1}` and id `{section_id_2}` swapped.")
 
 @section.command(name="delete")
-async def delete(ctx, section_id: int = -1):
+async def section_delete(ctx, section_id: Int__AlertInChat("Invalid section id")):
+    if global_.int_conversion_error:
+        print("Bad integer input, returning")
+        return
     global_.platform[global_.branch][section_id].delete()
     await ctx.send(f"Section `{section_id}` in branch `{global_.branch}` deleted.")
 
 @section.command(name="display")
-async def display(ctx, section_id: int = -1):
-    if section_id == -1:
+async def section_display(ctx, section_id: Int__AlertInChat("Invalid section id")):
+    if global_.int_conversion_error:
+        print("Bad integer input")
         first_line = f"Listing {len(global_.platform[global_.branch])} policies"
         line_limit_exceeded = ", limited to first 30" if len(global_.platform[global_.branch]) > 30 else ''
         line = "-" * int((len(first_line) + len(line_limit_exceeded)) * 1.4)
@@ -127,7 +189,10 @@ async def display(ctx, section_id: int = -1):
             {len(global_.platform[global_.branch][section_id].policies_by_id)} policies.")
 
 @section.command(name="select")
-async def select(ctx, section_id: int = -1):
+async def section_select(ctx, section_id: Int__AlertInChat("Invalid section id")):
+    if global_.int_conversion_error:
+        print("Bad integer input, returning")
+        return
     global_.section = section_id
     await ctx.send(f"Set section selector to `{section_id}`.")
